Refactoring: Move scrub_json_payload() to og.api.validation:
This way we can reuse it in other API endpoints - it contains nothing that
is specific to webactions, and should be easy to extend further if needed.
